+++ src/Render.cpp
@@ -6,7 +6,7 @@ void draw()
 
 	debug();
 	
-	for (int i = 0; i < object_vec.size(); ++i)
+	for (size_t i = 0; i < object_vec.size(); ++i)
 	{
 		object_vec[i]->draw();
 	}
@@ -26,7 +26,7 @@ void debug()
 {
 	if (direction)
 	{
-		for (int i = 0; i < object_vec.size(); ++i)
+		for (size_t i = 0; i < object_vec.size(); ++i)
 		{
 			static_cast<Circle&>(*object_vec[i]).debug();
 		}
